+++ src/SBP_coefficients/Mattsson2014.jl
@@ -313,7 +313,7 @@ function second_derivative_coefficients(source::Mattsson2014, order::Int, T = Fl
         right_weights = left_weights
         left_boundary_derivatives = (
                                      # first derivative
-                                     DerivativeCoefficientRow{T, 1, 4}(SVector(T(-11 / 6),
+                                     DerivativeCoefficientRow{T, 1, 4}(SVector(T(-11 // 6),
                                                                                T(3),
                                                                                T(-3 // 2),
                                                                                T(1 // 3))),)
@@ -446,7 +446,7 @@ function second_derivative_coefficients(source::Mattsson2014, order::Int, T = Fl
         right_weights = left_weights
         left_boundary_derivatives = (
                                      # first derivative
-                                     DerivativeCoefficientRow{T, 1, 5}(SVector(T(-25 / 12),
+                                     DerivativeCoefficientRow{T, 1, 5}(SVector(T(-25 // 12),
                                                                                T(4),
                                                                                T(-3),
                                                                                T(4 // 3),
@@ -568,7 +568,7 @@ function third_derivative_coefficients(source::Mattsson2014, order::Int, T = Flo
         right_weights = left_weights
         left_boundary_derivatives = (
                                      # first derivative
-                                     DerivativeCoefficientRow{T, 1, 4}(SVector(T(-11 / 6),
+                                     DerivativeCoefficientRow{T, 1, 4}(SVector(T(-11 // 6),
                                                                                T(3),
                                                                                T(-3 // 2),
                                                                                T(1 // 3))),
@@ -863,7 +863,7 @@ function fourth_derivative_coefficients(source::Mattsson2014, order::Int, T = Fl
         right_weights = left_weights
         left_boundary_derivatives = (
                                      # first derivative
-                                     DerivativeCoefficientRow{T, 1, 4}(SVector(T(-11 / 6),
+                                     DerivativeCoefficientRow{T, 1, 4}(SVector(T(-11 // 6),
                                                                                T(3),
                                                                                T(-3 // 2),
                                                                                T(1 // 3))),